requirements.txt
aioredis==1.3.1
amqp==2.6.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
click==7.1.2
configparser==5.0.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.7
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
future==0.18.2
gunicorn==20.0.4
h11==0.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
kombu==4.6.10
llvmlite==0.33.0
msgpack==0.6.2
mss==5.1.0
numba==0.50.0
numpy==1.18.5
opencv-contrib-python==4.2.0.34
opencv-contrib-python-headless==4.2.0.34
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
Pillow==7.1.2
prettyconf==2.1.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
Unipath==1.1
uvicorn==0.11.5
vine==1.3.0
websockets==8.1
whitenoise==5.1.0
wincertstore==0.2
zipp==3.1.0
zope.interface==5.1.0
